+++ sdk/iot/Azure.Iot.Hub.Service/samples/IotHubClientSamples/BulkDeviceIdentityLifecycleSamples.cs
@@ -98,6 +98,8 @@ namespace Azure.Iot.Hub.Service.Samples
             }
             catch (Exception ex)
             {
+                // Try to cleanup before exiting with fatal error.
+                await CleanupHelper.DeleteAllDevicesInHubAsync(IoTHubServiceClient);
                 SampleLogger.FatalError($"Failed to create device identity due to:\n{ex}");
                 throw;
             }
@@ -143,6 +145,8 @@ namespace Azure.Iot.Hub.Service.Samples
             }
             catch (Exception ex)
             {
+                // Try to cleanup before exiting with fatal error.
+                await CleanupHelper.DeleteAllDevicesInHubAsync(IoTHubServiceClient);
                 SampleLogger.FatalError($"Failed to update a device identity due to:\n{ex}");
                 throw;
             }
@@ -189,6 +193,8 @@ namespace Azure.Iot.Hub.Service.Samples
             }
             catch (Exception ex)
             {
+                // Try to cleanup before exiting with fatal error.
+                await CleanupHelper.DeleteAllDevicesInHubAsync(IoTHubServiceClient);
                 SampleLogger.FatalError($"Failed to update a device identity due to:\n{ex}");
                 throw;
             }
@@ -229,6 +235,8 @@ namespace Azure.Iot.Hub.Service.Samples
             }
             catch (Exception ex)
             {
+                // Try to cleanup before exiting with fatal error.
+                await CleanupHelper.DeleteAllDevicesInHubAsync(IoTHubServiceClient);
                 SampleLogger.FatalError($"Failed to device identity due to:\n{ex}");
             }
         }