+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/src/TelemetryType.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace OpenTelemetry.Exporter.AzureMonitor
+{
+    internal enum TelemetryType
+    {
+        Request,
+        Dependency,
+        Message,
+        Event
+    }
+}