+++ sdk/storage/Azure.Storage.Blobs.ChangeFeed/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
-## 12.0.0-preview.4 (Unreleased)
+## 12.0.0-preview.5 (Unreleased)
 
+## 12.0.0-preview.4 (2020-08-18)
+- Bug in TaskExtensions.EnsureCompleted method that causes it to unconditionally throw an exception in the environments with synchronization context
 
 ## 12.0.0-preview.3 (2020-08-13)
 - This release contains bug fixes to improve quality.